+++ reptar/__init__.py
@@ -6,7 +6,6 @@ __version__ = _version.get_versions()["version"]
 
 from .creating import Creator
 from .reptar_file import File
-from .save import Saver
 from .sampling import Sampler
 
-__all__ = ["Creator", "File", "Saver", "Sampler"]
+__all__ = ["Creator", "File", "Sampler"]